线上问题记录bug修复

+++ armor-api/src/main/java/com/zhixinhuixue/armor/model/dto/request/OnlineQuestionReqDTO.java
@@ -30,7 +30,7 @@ public class OnlineQuestionReqDTO {
     private Date endTime;
 
     @NotNull(message = "工作时间不能为空")
-    private Integer workHour;
+    private Float workHour;
 
     private List<String> urlList;
 
@@ -74,11 +74,11 @@ public class OnlineQuestionReqDTO {
         this.endTime = endTime;
     }
 
-    public Integer getWorkHour() {
+    public Float getWorkHour() {
         return workHour;
     }
 
-    public void setWorkHour(Integer workHour) {
+    public void setWorkHour(Float workHour) {
         this.workHour = workHour;
     }
